ticket 8 + ticket 10
Ability to set the current shunt resistor, max current and max voltage of the board and have it check that it does exceed the bounds of the device.
For the calibration function, use the provided information to confirm that the device will be configured correctly and perform an automatic calibration.