implement unit test and modifying test classes
implement some of unit test. each test checks the operation of the system when there is interaction between the objects. for instance: Checking the change of timer when a keypress event occurs, Checking the change of menu display when a keypress event occurs or having a change from 'INA219' module. 
besides, we changed kind of inheritance of classes from private to protect . so, now unit test can be modified separately .